% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{altitude}
\alias{altitude}
\alias{altitude.dive}
\alias{altitude.ndive}
\title{altitude}
\usage{
altitude(object)

\method{altitude}{dive}(object)

\method{altitude}{ndive}(object)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} and \code{\link[DiveR]{ndive}} objects.}
}
\value{
It returns a numeric with the altitude of the dive. 
Is a vector if working on \code{\link[DiveR]{ndive}} object
}
\description{
\code{altitude} retrieve the altitude of a singular or multiple dive 
sequence.
}
\examples{
# Simple dive
altitude(dive(20,40, altitude = 1500))
# Multiple dives
altitude(ndive(dive(20,40, altitude = 1000), dive(15, 80), inter = 540))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
