% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{altitude_depth}
\alias{altitude_depth}
\title{altitude_depth}
\usage{
altitude_depth(depth, altitude = 0)
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.}

\item{altitude}{Heigth in meter from sea level, it will impact desaturation
process and ascen_speed. Default is sea level (0m).}
}
\value{
The equivalent depth in meter when diving at altitude.
}
\description{
Compute equivalent depth when diving at altitude.
}
\details{
Computation is below :
Equivalent depth = depth * (sea level pressure / altitude pressure)

Altitude pressure is defines in \code{\link[DiveR]{altitude_pressure}}

Sea level pressure = 1 bar.

Value is rounded to upper value (so to lower equivalent depth).
 
 @examples 
 altitude_depth(depth = 20)
 altitude_depth(depth = 24, altitude = 2000)
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
