% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{altitude_pressure}
\alias{altitude_pressure}
\title{altitude_depth}
\usage{
altitude_pressure(altitude = 0)
}
\arguments{
\item{altitude}{Heigth in meter from sea level, it will impact desaturation
process and ascen_speed. Default is sea level (0m).}
}
\value{
The equivalent depth in meter when diving at altitude.
}
\description{
Compute atmospherique pressure at altitude.
}
\details{
Altitude pressure is :
sea level pressure - ( altitude (m) / 1e4 )

Sea level pressure = 1 bar.

Value is rounded to upper value (so to lower equivalent depth).
 
 @examples 
 altitude_pressure(0)
 altitude_pressure(1000)
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
