% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_nitrox.R
\name{as.gas}
\alias{as.gas}
\alias{as.gas.gas}
\alias{as.gas.numeric}
\alias{as.gas.character}
\title{Coerce to gas}
\usage{
as.gas(x)

\method{as.gas}{gas}(x)

\method{as.gas}{numeric}(x)

\method{as.gas}{character}(x)
}
\arguments{
\item{x}{R object in numeric,  character}
}
\description{
Coerce to gas
}
\details{
Will call gas function, so names and will be 
deducted from the ppo2 as "NXppo2".
}
\examples{
as.gas(NULL)

as.gas(gas())

as.gas(c(ppo2 = 0.3, ppn2 = 0.7))

as.gas("AIR")
as.gas("EAN32")
as.gas("NX36")

}
