% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beepodyna.R
\docType{package}
\name{beepodyna}
\alias{beepodyna}
\alias{beepodyna-package}
\title{BeePODYNA}
\usage{
beepodyna(label, community,
  interactions = interactions(length(community)), functions = identity,
  verbose = TRUE)
}
\arguments{
\item{label}{is the label of the \code{beepodyna} object.}

\item{community}{is the community object of class \code{community} to study}

\item{interactions}{is the interaction matrix of the class \code{interactions} between the populations of the community. Default is a matrix without interactions.}

\item{functions}{is a function or a list of functions to apply to each population in the same order as it is in the community.
If the length of the list is not the same as the community, the functions are repeated so the length of the list is the same as the number of populations.}

\item{verbose}{Default is \code{TRUE}. Set to \code{FALSE} if you don't want the warnings.}
}
\description{
A package to study population dynamic.

The \code{beepodyna} function creates a \code{R} object of the class \code{beepodyna}.
}
\details{
The \code{beepodyna} object is a list containing a community, the corresponding interaction matrix
and a list of the functions to apply to each population, in the order they appear in the community object.
}
\examples{
hare = population("hirsuta",30,2,80)
lynx = population("daonensis",4,1.2,60)
hudson = community("hudson",hare,lynx)

beepodyna(label="model_1",
          community=hudson,
          interactions=interactions(2),
          exp,
          verbose = FALSE
          )

}
\seealso{
\code{\link[BeePODYNA]{community}} and \code{\link[BeePODYNA]{interactions}} to create the object in parameters.
}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
