% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ceiling_dec}
\alias{ceiling_dec}
\title{ceiling_dec}
\usage{
ceiling_dec(x, level = 1)
}
\arguments{
\item{x}{values to round. Must be numeric}

\item{level}{decimal to round to. Default is 1}
}
\value{
data.frame
}
\description{
Ceiling at a specific decimal level
}
\details{
copied from SO
\url{https://stackoverflow.com/questions/35807523/r-decimal-ceiling}
}
\examples{
ceiling_dec(1.259, 2)
ceiling_dec(1.9, 0)
ceiling_dec(29, -1)
ceiling_dec(1.251, 2)
ceiling_dec(1.2, 0)
ceiling_dec(23, -1)

}
