% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso.R
\name{conso}
\alias{conso}
\title{conso}
\usage{
conso(dive, tank, cons = 20, failure_label = "AF")
}
\arguments{
\item{dive}{\code{\link[DiveR]{dive}} object}

\item{tank}{\code{\link[DiveR]{tank}} object or a list of tank objects.
Priority of consumption for tanks is set by their order in list.}

\item{cons}{Litre per minute breathed by diver. Single numeric positive value.
20 L/min by default}

\item{failure_label}{Label for when a tank is a empty. Single character
string. 'AF' by default.}
}
\value{
conso, a conso class object.
}
\description{
conso
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
