% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco.R
\name{curve}
\alias{curve}
\title{curve}
\usage{
curve(time, depth, palier, vup = 10)
}
\arguments{
\item{time}{in minute. Max is 180' see Detail !}

\item{depth}{in meter. Max is 65m see Detail !}

\item{palier}{a object palier computed with the palier function. 
Is a list of depth and time for every deco stage.}

\item{vup}{10 m/min max speed up to the deco. 
Speed between deco is fixed to 6}
}
\value{
a list, with a vector of depths, a vector of time and the dtr value.
}
\description{
compute the time from end of dive to surface
}
\details{
This function will stop if the depth > 65 or time > 180 because the table are limited to this extent.
However for lower values the table can return NA values.
}
