% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{cust_axis}
\alias{cust_axis}
\title{cust_axis}
\usage{
cust_axis(dive, col)
}
\arguments{
\item{dive}{a \code{\link[mn90]{dive}} object.}

\item{col}{a color}
}
\description{
draw the xaxis with time input
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
