% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_haldane.R
\name{cut_dtcurve}
\alias{cut_dtcurve}
\title{cut_dtcurve}
\usage{
cut_dtcurve(dtcurve, delta = 1, cut_h = FALSE)
}
\arguments{
\item{dtcurve}{a depth time curve in a data.frame with 2 columns depth and 
time. Depths values are meters (positive values) and time is in minute.}

\item{delta}{time between 2 points. Default value is set to 1 minute, 
minor value will add computation but larger value can be dangerous and 
approximate desaturation.}

\item{cut_h}{is the algorithme going to cut horizontal part of the dtcurve. 
FALSE by default.}
}
\value{
discretised dtcurve table.
}
\description{
Adding points in a dtcurve to compute more precise desaturation.
}
