% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R, R/methods_conso.R,
%   R/methods_dive.R
\name{depth}
\alias{depth}
\alias{depth.dive}
\alias{depth.ndive}
\alias{depth.conso}
\title{depth}
\usage{
depth(object)

\method{depth}{dive}(object)

\method{depth}{ndive}(object)

\method{depth}{conso}(object)

depth(object)

\method{depth}{dive}(object)

\method{depth}{ndive}(object)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}}, \code{\link[DiveR]{ndive}} and
 \code{\link[DiveR]{conso}} objects.}
}
\value{
It returns a numeric with the depth of the dive. 
Is a vector if working on ndive object

It returns a numeric with the depth of the dive. 
Is a vector if working on \code{\link[DiveR]{ndive}} object
}
\description{
\code{depth} retrieve the maximum depth of a singular or multiple dive sequence.

\code{depth} retrieve the maximum depth of a singular or multiple dive 
sequence. As conso has a dive inside, the depth function also works on it.
}
\examples{
# Simple dive
depth(dive(20,40))
# Multiple dives
depth(ndive(dive(20,40), dive(15, 80), inter = 540))

# Simple dive
depth(dive(20,40))
# Multiple dives
depth(ndive(dive(20,40), dive(15, 80), inter = 540))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
