% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_haldane.R
\name{desat_haldane}
\alias{desat_haldane}
\title{desat_haldane}
\usage{
desat_haldane(dtcurve, maj = 0, altitude = 0, ppn2 = 0.791, ncomp = 5)
}
\arguments{
\item{dtcurve}{a depth time curve in a data.frame with 2 columns depth and 
time. Depths values are meters (positive values) and time is in minute.}

\item{maj}{TODO : ignored parameter for now}

\item{altitude}{Altitude of the dive in meter. Default is 0 m (sea level).}

\item{ppn2}{Partial pressure of nitrogen in bar. Default is 0.791 bar}

\item{ncomp}{Number of compartment for the model. Choice between 5 and 12.}
}
\value{
a desat object, which is a list with a data.frame containing 
desaturation stops at 9, 6 and 3 m depth. Next element is the dive group
for possible second dive and lastly the times at which the desaturation
stops occur during the dive. The last element is NULL because it's made with 
tables.
}
\description{
desat_haldane
}
\details{
There is a modification of the last compartment, with an half-life
of 60 instead of 75 min. This is because the Sc value was not found for such
time.

Limitation is imposed because we use Haldane model with constant depths,
and not the Schreiner equation using depth evolution.
}
