% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_utils.R
\name{dist2time}
\alias{dist2time}
\title{dist2time}
\usage{
dist2time(dist, speed)
}
\arguments{
\item{dist}{Distance vector in meters from points to points.}

\item{speed}{A diver speed value in meter/minute.}
}
\description{
dist2time
}
