% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiveR.R
\name{dive}
\alias{dive}
\title{dive}
\usage{
dive(
  depth = 20,
  time = 40,
  secu = TRUE,
  vup = 10,
  maj = 0,
  hour = NULL,
  dist = NULL,
  speed = NULL,
  way = c("AS", "AR")
)
}
\arguments{
\item{depth}{in meter}

\item{time}{in minute}

\item{secu}{true by default, secu deco stage 3 min at 3 meter}

\item{vup}{10 m/min by default}

\item{maj}{0 by default}

\item{hour}{NULL not implemented yet}

\item{dist}{a distance vector}

\item{speed}{speed of the diver}

\item{way}{if the dive is one way or the diver return by the same depth}
}
\value{
dive, a dive class object.
}
\description{
dive
}
\details{
See \code{\link[DiveR]{tablecheck}} for limit values of depth and time.
}
\examples{
dive = dive(depth = 39, time = 22, secu = TRUE, vup = 10)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
