% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diveR_packages.R
\name{diveR_packages}
\alias{diveR_packages}
\title{List all packages in diveR}
\usage{
diveR_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include diveR in the list?}
}
\description{
List all packages in diveR
}
\examples{
diveR_packages()
}
