% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{dtime}
\alias{dtime}
\alias{dtime.dive}
\title{dtime}
\usage{
dtime(object)

## S3 method for class 'dive'
# dtime(dive)
}
\arguments{
\item{object}{is a mn90 object. There are methods for 
\code{\link[mn90]{dive}} objects.}
}
\value{
a single numeric value
}
\description{
\code{dtime} retrive the depth time of a singular or multiple dive sequence.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
