% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(tank, dive)
}
\arguments{
\item{tank}{\code{\link[DiveR]{tank}} object or a list of tank objects.
Priority of consumption for tanks is set by their order in list.}

\item{dive}{\code{\link[DiveR]{dive}} object}
}
\description{
expand
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
