% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_nitrox.R
\name{gas}
\alias{gas}
\title{Set a gas for dive saturation}
\usage{
gas(ppo2 = 0.209, ppn2 = 0.791, name = character(0))
}
\arguments{
\item{ppo2}{partial percentage of dioxygen in breathed mixture. in percentage}

\item{ppn2}{partial percentage of nitrogen in breathed mixture. in percentage}

\item{name}{default is empty and will be deducted from the ppo2 as "NXppo2". 
This is equivalent to "EANppo2"}
}
\description{
Gas for dive. The values are taken into account for tissue saturation models.
}
\details{
Sum of ppn2 and ppo2 must be 1. There is two cases :
if the sum is superior to 0, the function will trigger an error; 
if the sum is inferior to 1, a gas "other" will compensate and the function
will only throw a warning.
}
\examples{
gas()
gas(0.32, 0.68, name = "NX32")
gas(0.32, 0.68)

}
