% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{grate_population}
\alias{grate_population}
\title{grate_population}
\usage{
grate_population(pop)
}
\arguments{
\item{pop}{is an object of the class 'population'.}
}
\value{
It returns the growth rate of the population.
}
\description{
\code{grate_population} returns the size of the population given in parameter.
}
\examples{
    pop<-population("Pop_1",42,2,200)
    grate_population(pop)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
