% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{growth_rate}
\alias{growth_rate}
\title{Calculates a growth rate from a birth rate and a death rate.}
\usage{
growth_rate(birth_rate, death_rate)
}
\arguments{
\item{birth_rate}{the birth rate of the population}

\item{death_rate}{the death rate of the population}
}
\value{
returns the growth rate of the population
}
\description{
Calculates a growth rate from a birth rate and a death rate.
}
\examples{
  growth_rate(0.5, 0.2)

}
