% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{grp}
\alias{grp}
\title{MN90 table}
\format{
An object of class \code{data.frame} with 25 rows and 27 columns.
}
\source{
table beuchat
}
\usage{
grp
}
\description{
A dataset containing an array with the palier time depending on depth and time and a data.frame with the group.
}
\author{
Jaunatre Maxime <maxime.jaunatre@gowachin974.fr>
}
\keyword{datasets}
