% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Haldane5}
\alias{Haldane5}
\alias{Haldane12}
\title{Haldane S values}
\format{
Each data.frame has a column 'comp' for the half-life of the compartment, 
and a column 'Scomp' for the S value of the compartment. For a data.frame
of HaldaneN there is N rows.

An object of class \code{data.frame} with 12 rows and 2 columns.
}
\source{

}
\usage{
Haldane5

Haldane12
}
\description{
A dataset containing different data.frames for Haldane S values.
}
\details{
\itemize{
 \item{Haldane5 - Compartments from 5 to 60 minutes}
 \item{Haldane12 - Compartments from 5 to 120 minutes}
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
\keyword{datasets}
