% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/desat_haldane.R
\name{cpp_half_life}
\alias{cpp_half_life}
\alias{half_life}
\title{half_life}
\usage{
cpp_half_life(period, time)

half_life(period, time)
}
\arguments{
\item{period}{period in minute for the compartment.}

\item{time}{time at which user want the load of nitrogen}
}
\value{
percentage load.
}
\description{
compute the load of nitrogen at a given time for a compartment period.
}
\examples{
cpp_half_life(period = 1, time = c(1,2))

half_life(50, 50)
cpp_half_life(period = 1, time = c(1,2))


}
