% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_utils.R
\name{init_dtcurve}
\alias{init_dtcurve}
\title{init_dtcurve}
\usage{
init_dtcurve(depth, time, ascent_speed = 10, way = c("OW", "WB"))
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.
A single value will result in square dive.}

\item{time}{Duration of the dive in minute. Need to be positive values.
A single value will result in square dive.}

\item{ascent_speed}{Ascent_speed in meter/minute. 10 m/min by default. 
Most dive table advice to limite this speed to 20M/min maximum.}

\item{way}{If the dive is one way (one way : 'OW') or if the diver return by 
the same depth (way back : 'WB').}
}
\description{
expand depth and time to a dtcurve data.frame
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
