% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insertRow}
\alias{insertRow}
\title{insertRow}
\usage{
insertRow(existingDF, newrow, r)
}
\arguments{
\item{existingDF}{The data.frame to insert row in.}

\item{newrow}{row to insert. Must be of length == ncol(existingDF)}

\item{r}{place where to insert the row}
}
\value{
data.frame
}
\description{
Insert a row at a row number. This an intern function.
}
\details{
copied from SO
\url{https://stackoverflow.com/questions/11561856/add-new-row-to-dataframe-at-specific-row-index-not-appended}
}
\examples{
existingDF <- as.data.frame(matrix(seq(20),nrow=5,ncol=4))
insertRow(existingDF, newrow = seq(4), r = 3)

}
