% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_conso.R
\name{is.conso}
\alias{is.conso}
\title{is.conso}
\usage{
is.conso(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE or FALSE
}
\description{
check if the object is of \code{\link[DiveR]{conso}} class
}
\examples{
t <- tank(12,200)
d <- dive(20, 40)
c <- conso(d, t)
is.tank(c)
is.tank('conso')

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
