% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_table.R
\name{is.desat}
\alias{is.desat}
\title{is.desat}
\usage{
is.desat(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE or FALSE
}
\description{
check if the object is of \code{\link[DiveR]{desat_table}} class
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
