% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{is.dive}
\alias{is.dive}
\title{is.dive}
\usage{
is.dive(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE or FALSE
}
\description{
check if the object is of \code{\link[DiveR]{dive}} class
}
\examples{
d <- dive(20,40)
is.dive(d)
is.dive('dive')

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
