% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_conso.R
\name{is.tank}
\alias{is.tank}
\title{is.tank}
\usage{
is.tank(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE or FALSE
}
\description{
check if the object is of \code{\link[DiveR]{tank}} class
}
\examples{
t <- tank(12,200)
is.tank(t)
is.tank('tank')

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
