% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{is_community}
\alias{is_community}
\title{is_community}
\usage{
is_community(x)
}
\arguments{
\item{x}{the object which must be a community to validate the condition}
}
\value{
a logical "TRUE" or "FALSE"
}
\description{
Check if the object is a community or not.
}
\examples{
hare = population("hirsuta",30,2,80)
lynx = population("daonensis",4,1.2,60)

hudson = community('hudson',hare,lynx)
is_community(hudson)

}
\seealso{
\code{\link[BeePODYNA]{community}} to see how to make an object of class community.
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
