% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{is_interactions}
\alias{is_interactions}
\title{is_interactions}
\usage{
is_interactions(x)
}
\arguments{
\item{x}{is an \code{R} object.}
}
\description{
\code{is_interactions} returns \code{TRUE} if x is an interaction matrix of the class 'interactions'.
It returns \code{FALSE} otherwise.
To create such an object you can use the \code{\link[BeePODYNA]{interactions}} function.
}
\examples{
  is_interactions(3)
  x=interactions(nb_pop=3, list.interactions=c(0.2,-0.5, 0.1,0.2,0.3,0.8))
  is_interactions(x)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
