% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_population.R
\name{is_population}
\alias{is_population}
\title{is_population}
\usage{
is_population(x)
}
\arguments{
\item{x}{the object which must be a population to validate the condition}
}
\value{
a logical "TRUE" or "FALSE"
}
\description{
The function determinines if the object is a population or not
}
\author{
Cresciense Lecaude
}
