% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{label}
\alias{label}
\alias{label.population}
\alias{label.community}
\alias{label.beepodyna}
\title{Returns the label of a BeePODYNA object}
\usage{
label(object)

\method{label}{population}(object)

\method{label}{community}(object)

\method{label}{beepodyna}(object)
}
\arguments{
\item{object}{is a BeePODYNA object.}
}
\value{
It returns a character with the label of the object.
}
\description{
\code{label} returns the label of \code{\link[BeePODYNA]{population}},
\code{\link[BeePODYNA]{community}} or \code{\link[BeePODYNA]{beepodyna}} object.
}
\examples{
    pop<-population("Pop_1",42,2,200)
    label(pop)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
