% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco.R
\name{majoration}
\alias{majoration}
\title{majoration}
\usage{
majoration(depth, group = "Z", inter = 16)
}
\arguments{
\item{depth}{in meter. Max is 65m see Detail !}

\item{group}{byt default "Z", the deco group indicated by a letter. 
This value is indicated in a palier object computed with the palier function.}

\item{inter}{16 by default, interval between dives in minutes}
}
\value{
palier, a list with a vector of depth and a vector of time. 
Vectors are ordered from deepest deco stage (9m) to the higher (3m).
}
\description{
compute the time majoration to a second dive at a specific depth.
}
\details{
This function will stop if the depth > 65 or time > 180 because the table 
are limited to this extent.
However for lower values the table can return NA values.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
