% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco.R
\name{max_depth_t}
\alias{max_depth_t}
\title{max_depth_t}
\usage{
max_depth_t(depth)
}
\arguments{
\item{depth}{in meter. Max is 65m see Detail !}
}
\value{
a single numeric value, the max time possible to dive at the given 
depth is the MN90 table.
}
\description{
Max time present in the table for a given depth.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
