% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{minesweeper}
\alias{minesweeper}
\title{Minesweeper}
\usage{
minesweeper(N = 20, X = 20, Y = 20, verbose = FALSE)
}
\arguments{
\item{N}{the number of mines. 20 by default.}

\item{X}{size of the matrix. 20 by default.}

\item{Y}{sive of the matrix. 20 by default.}

\item{verbose}{FALSE by default.}
}
\description{
Try to clear all paths without jumping on a mine.
You need to click on the cells to discover the plot round after round.
}
