% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrox.R
\name{MODcheck}
\alias{MODcheck}
\title{MODcheck}
\usage{
MODcheck(depth, ppn2 = 0.791, force = FALSE)
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.}

\item{ppn2}{Partial pressure of nitrogen in bar. Default is 0.791 bar}

\item{force}{FALSE by default, if TRUE don't stop the function but 
return a TRUE/FALSE value}
}
\value{
Nothing expect when \code{force} parameter is TRUE. 
Is designed to stop code and returns errors
}
\description{
Test if the depth is possible with such partial pressure of nitrogen.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
