% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrox.R
\name{nitrox_depth}
\alias{nitrox_depth}
\title{nitrox_depth}
\usage{
nitrox_depth(depth, ppo2 = 0.209, MOD = 1.6)
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.}

\item{ppo2}{Partial pressure of dioxygen in bar. Default is 0.209 bar}

\item{MOD}{maximum operating depth. This is the maximal ppo2 value in bar
default is 1.6}
}
\value{
The equivalent depth in meter when using this nitrox partial pressure
in a tank.
}
\description{
Compute equivalent depth when using this nitrox partial pressure
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
