% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrox.R
\name{nitrox_maxdepth}
\alias{nitrox_maxdepth}
\title{nitrox_maxdepth}
\usage{
nitrox_maxdepth(ppn2 = 0.791)
}
\arguments{
\item{ppn2}{Partial pressure of nitrogen in bar. Default is 0.791 bar}
}
\value{
maximum depth in meter
}
\description{
Maximum depth for partial pressure of nitrogen.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
