% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco.R
\name{palier}
\alias{palier}
\title{palier}
\usage{
palier(depth, time, secu = TRUE)
}
\arguments{
\item{depth}{in meter. Max is 65m see Detail !}

\item{time}{in minute. Max is 180' see Detail !}

\item{secu}{true by default, add a secu deco 3 min at 3 meter}
}
\value{
palier, a list with a vector of depth and a vector of time. 
Vectors are ordered from deepest deco stage (9m) to the higher (3m).
}
\description{
compute the palier depth and time from mn90 table.
}
\details{
This function will stop if the depth > 65 or time > 180 because the table are limited to this extent.
However for lower values the table can return NA values.
}
