% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.community}
\alias{plot.community}
\title{plot.community}
\usage{
\method{plot}{community}(x, ...)
}
\arguments{
\item{x}{a population object}

\item{...}{every argument for the \code{\link[graphics]{plot}} function such as graphical parameters for lines.
Classical \code{\link[graphics]{graphical parameter}} applies on the population curve, NOT on the capacity line (custom parameter need to be given, as exemple capacity_lty).
\itemize{
  \item \strong{capacity_line} set to \code{TRUE} by default, if a dashed line must be drawned to show the capacity.
  \item \strong{capacity_lty} set to \code{3} by default.
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{text_print} set to \code{TRUE} by default, whether there is a text on the capacipty line or not (writting population 'K label' by default).
  \item \strong{text_x} set to \code{0} by default, where to position the capacity text on X axis, Y axis is the capacity value.
  \item \strong{xlab} set to \code{'Time'} by default, see \code{\link[graphics]{title}}.
  \item \strong{ylab} set to \code{'Size'} by default, see \code{\link[graphics]{title}}.
  \item \strong{main} set to use the pop label by default, see \code{\link[graphics]{title}}.
  }}
}
\description{
plot one given population and its capacity line.
}
\note{
For \strong{capacity_line}, \strong{capacity_lty}, \strong{type}, \strong{text_print} and \strong{text_x} parameters, values can be a single value or a vector of value. If the vector length is inferior to the population number in the community, the value or the vector will be copy.
}
\examples{
data(hudson)
plot(hudson)

}
\seealso{
\itemize{
\item \code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and \code{\link[graphics]{par}} for plot parameter that were omitted on this documentation
\item \code{\link[BeePODYNA]{plot.population}}, \code{\link[BeePODYNA]{population}}, \code{\link[BeePODYNA]{community}} for every aspect about population and community creations.
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
