% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{plot.conso}
\alias{plot.conso}
\title{plot.conso}
\usage{
\method{plot}{conso}(
  x,
  ...,
  dive_print = TRUE,
  rules_print = TRUE,
  hour_print = TRUE,
  line_print = TRUE,
  depth_print = TRUE,
  time_print = TRUE,
  def_cols = FALSE,
  legend = TRUE,
  add = FALSE
)
}
\arguments{
\item{x}{an object of class conso.}

\item{...}{every argument for the \code{\link[graphics]{plot}} function 
such as graphical parameters for lines.
Classical \code{\link[graphics]{graphical parameter}} applies on the 
dive curve.
\itemize{
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{xlab} set to \code{'Time'} by default, 
  see \code{\link[graphics]{title}}.
  \item \strong{ylab} set to \code{'Depth'} by default, 
  see \code{\link[graphics]{title}}.
  }}

\item{dive_print}{set to \code{TRUE} by default, whether there is the dive
curve printed.}

\item{rules_print}{set to \code{TRUE} by default, whether the rules are 
shown}

\item{hour_print}{set to \code{TRUE} by default, whether there is the hour
at beginning and end of the dive.}

\item{line_print}{set to \code{TRUE} by default, whether there is lines 
for depths.}

\item{depth_print}{set to \code{TRUE} by default, whether there is depth 
values along the dive.}

\item{time_print}{set to \code{TRUE} by default, whether there is time 
values along the dive.}

\item{def_cols}{set to \code{FALSE} by default, modify plot colors to match
shinyapp css.}

\item{legend}{set to \code{FALSE} by default, print legend colors for the 
tank.}

\item{add}{set to \code{FALSE} by default, to add another dive plot 
on a precedent one.}
}
\description{
Plot the dive curve and consumption depending on time and depth. 
Only represent sqare profile
}
\seealso{
\itemize{
\item \code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and 
\code{\link[graphics]{par}} for plot parameter that were omitted 
on this documentation
\item \code{\link[DiveR]{conso}} for every aspect about conso object 
creations.
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
