% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{plot.dive}
\alias{plot.dive}
\title{plot.dive}
\usage{
\method{plot}{dive}(x, ..., text_print = TRUE, add = FALSE)
}
\arguments{
\item{x}{an object of class dive.}

\item{...}{every argument for the \code{\link[graphics]{plot}} function 
such as graphical parameters for lines.
Classical \code{\link[graphics]{graphical parameter}} applies on the 
population curve, NOT on the capacity line (custom parameter need 
to be given, as exemple capacity_lty).
\itemize{
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{xlab} set to \code{'Time'} by default, 
  see \code{\link[graphics]{title}}.
  \item \strong{ylab} set to \code{'Size'} by default, 
  see \code{\link[graphics]{title}}.
  \item \strong{main} set to use the pop label by default, 
  see \code{\link[graphics]{title}}.
  }}

\item{text_print}{set to \code{TRUE} by default, whether there is a text 
on the capacipty line or not (writting population 'K label' by default).}

\item{add}{set to \code{FALSE} by default, to add another population plot 
on a precedent one.}
}
\description{
Plot the dive curve depending on time and depth. 
Only represent sqare profile
}
\seealso{
\itemize{
\item \code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and 
\code{\link[graphics]{par}} for plot parameter that were omitted 
on this documentation
\item \code{\link[mn90]{dive}} for every aspect about dive object creations.
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
