% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.population}
\alias{plot.population}
\title{plot_population}
\usage{
\method{plot}{population}(x, ..., capacity_line = TRUE,
  capacity_lty = 3, text_print = TRUE, add = FALSE)
}
\arguments{
\item{x}{a population object}

\item{...}{every argument for the \code{\link[graphics]{plot}} function such as graphical parameters for lines.
Classical \code{\link[graphics]{graphical parameter}} applies on the population curve, NOT on the capacity line (custom parameter need to be given, as exemple capacity_lty).
\itemize{
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{text_x} set to \code{0} by default, where to position the capacity text on X axis, Y axis is the capacity value.
  \item \strong{xlab} set to \code{'Time'} by default, see \code{\link[graphics]{title}}.
  \item \strong{ylab} set to \code{'Size'} by default, see \code{\link[graphics]{title}}.
  \item \strong{main} set to use the pop label by default, see \code{\link[graphics]{title}}.
  }}

\item{capacity_line}{set to \code{TRUE} by default, if a dashed line must be drawned to show the capacity.}

\item{capacity_lty}{set to \code{3} by default.}

\item{text_print}{set to \code{TRUE} by default, whether there is a text on the capacipty line or not (writting population 'K label' by default).}

\item{add}{set to \code{FALSE} by default, to add another population plot on a precedent one.}
}
\description{
plot one given population and its capacity line.
}
\examples{
library(Template)
data(hudson)
hare <- hudson$hare
lynx <- hudson$lynx

plot(hare, col = 'red', pch=15, type = 'b', capacity_lty = 3,
     lty =2, main = "Hudson data")
plot(lynx, col = 'blue', pch=15, type = 'b', capacity_lty = 3,
     add = TRUE, lty = 1, capacity_line = TRUE)

}
\seealso{
\itemize{
\item \code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and \code{\link[graphics]{par}} for plot parameter that were omitted on this documentation
\item \code{\link[BeePODYNA]{population}} for every aspect about population and community creations.
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
