% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{ppo2}
\alias{ppo2}
\alias{ppo2.dive}
\alias{ppo2.ndive}
\title{ppo2}
\usage{
ppo2(object)

\method{ppo2}{dive}(object)

\method{ppo2}{ndive}(object)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} objects.}
}
\value{
a single numeric value
}
\description{
\code{ppo2} retrieve the partial pressure of O2 of a singular or 
multiple dive sequence.
}
\examples{
# Simple dive
ppo2(dive(20,40))
# Multiple dives
ppo2(ndive(dive(20,40), dive(15, 80), inter = 540))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
