% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_conso.R
\name{pressure}
\alias{pressure}
\alias{pressure.tank}
\alias{pressure.conso}
\title{pressure
\code{pressure} retrieve the pressure of tank(s).}
\usage{
pressure(object, ...)

\method{pressure}{tank}(object, ...)

\method{pressure}{conso}(object, ..., hour = FALSE)
}
\arguments{
\item{object}{is a DiveR object. There is methods for 
\code{\link[DiveR]{tank}} and \code{\link[DiveR]{conso}} objects.}

\item{...}{other parameters passed to the pressure.conso method.
\describe{
  \item{tankn}{the number of the tank if there is multiple. NULL by default
  will return all the tanks}
  \item{time}{a time in minute. NULL by default will get the last pressure}
  \item{hour}{}
}}

\item{hour}{is the time defined inside the hours of the dive (TRUE) or 
in relative time to dive start (FALSE). Set to FALSE by default.}
}
\value{
single numeric value for \code{\link[DiveR]{tank}} object and vector for 
 \code{\link[DiveR]{conso}} object. 
 
 @examples 
 t <- tank(12,200)
 pressure(object = t)
 
 d <- dive(20,15)
 c <- conso(d, t, cons = 20)
 pressure(object = c)
 
 pressure(object = c, time = 7)
}
\description{
pressure
\code{pressure} retrieve the pressure of tank(s).
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
