% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{rm_secu}
\alias{rm_secu}
\title{rm_secu}
\usage{
rm_secu(dive)
}
\arguments{
\item{dive}{\code{\link[DiveR]{dive}} object.}
}
\description{
Remove the last security desaturation stop (3min at 3m)
}
\examples{
d1 <- dive(20, 40)
d2 <- rm_secu(d1)
plot(d1)
plot(d2, add = TRUE, col = "darkred")

# When there is other desat stops
d1 <- dive(39, 22)
d2 <- rm_secu(d1)
plot(d1)
plot(d2, add = TRUE, col = "darkred")

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
