% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{secu}
\alias{secu}
\alias{secu.dive}
\title{secu}
\usage{
secu(object)

## S3 method for class 'dive'
# depth(dive)
}
\arguments{
\item{object}{is a mn90 object. There are methods for 
\code{\link[mn90]{dive}} objects.}
}
\value{
Boolean
}
\description{
\code{secu} retrieve if a dive was set with secu = TRUE or FALSE.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
