% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn90.R
\name{shiny_mn90_app}
\alias{shiny_mn90_app}
\title{mn90 app}
\usage{
shiny_mn90_app()
}
\description{
custom function to run the shiny app (thinkR idea, to explore in docker)
}
\examples{
mn90::shiny_mn90_app()

}
