% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{simpl}
\alias{simpl}
\title{simpl}
\usage{
simpl(dtcurve)
}
\arguments{
\item{dtcurve}{a curve dive list}
}
\description{
Remove points of curve when isodepth around
}
\examples{
cons <- conso(dive = dive(20,40), bloc = bloc(10, 230), cons = 20, mid = 100, reserve = 50)
cons$dtcurve
simpl(cons$dtcurve)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
