% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{speed}
\alias{speed}
\alias{speed.dive}
\title{speed}
\usage{
speed(object, ...)

## S3 method for class 'dive'
# speed(dive)
}
\arguments{
\item{object}{is a mn90 object. There are methods for 
\code{\link[mn90]{dive}} objects.}
}
\value{
It returns a numeric with the duration of the ascent and deco.
}
\description{
\code{speed} compute the ascent speed from a depth and between deco stages 
of a dive.
It can be retrieved from a dive object.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
