% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{table}
\alias{table}
\title{MN90 table}
\format{
A community with 2 populations:
\describe{
  \item{grp}{deco groupe}
  \item{table}{palier table at depth of 3, 6 and 9 meters.}
  \itemize{
  \item{m3}
  \item{m6}
  \item{m9}
  }
}
}
\source{
table beuchat
}
\usage{
table
}
\description{
A dataset containing an array with the palier time depending on depth and time and a data.frame with the group.
}
\author{
Jaunatre Maxime <maxime.jaunatre@gowachin974.fr>
}
\keyword{datasets}
