% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_table.R
\name{table_ndive}
\alias{table_ndive}
\title{table_ndive}
\usage{
table_ndive(dive1, dive2, inter = 16, verbose = FALSE)
}
\arguments{
\item{dive1}{the first dive, must be a \code{\link[DiveR]{dive}} object}

\item{dive2}{the second dive, must be a \code{\link[DiveR]{dive}} object. 
This one will be modified with a majoration obtained from dive1 and 
the interval.}

\item{inter}{16 by default, interval in minute between the end of the first 
dive and the beginning of the second.}

\item{verbose}{allow cat return in consol for debug purposes. Show which
case of sequence is used.}
}
\value{
ndive, a ndive class object.
}
\description{
table_ndive
}
\details{
See \code{\link[DiveR]{tablecheck}} for limit values of depth and time 
of a dive.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
