% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso.R
\name{tank}
\alias{tank}
\title{tank}
\usage{
tank(
  vol,
  press,
  rules = list(rules = c(mid = 50, res = 25), sys = "\%"),
  gas = c("Air"),
  typ = c("back", "relay"),
  limit = NULL,
  name = NULL
)
}
\arguments{
\item{vol}{tank volume in litre.}

\item{press}{tank pression in bar.}

\item{rules}{tank rules to watch during a dive. A list of two named element :
\describe{
  \item{"rules"}{vector of 2 named integer indicating a percentage 
  or a pression. The names will be used in the plot function later}
  \item{"sys"}{character string, either '%' or 'bar'. Percentage must be
  between 0 and 100.}
}}

\item{gas}{tank gas, by default "Air". Parameter is here for future dev.}

\item{typ}{tank type, by default "back"
\describe{
  \item{"solo"}{single tank}
  \item{"relay"}{single tank to be dropped at certain time}
}}

\item{limit}{a two element vector with times between which the tank 
is not used. Can be used to mimic an accident, or a relay tank.}

\item{name}{Possibility to name the tank for better understanding after.
by default will be named after the typ and volume.}
}
\description{
Creation of tank object for usage in consumption part of a dive.
}
\details{
To set a relay tank, rule1 and rule2 must be the same. Therefore the tank 
won't be usable once pressure reach rule2 and until all other tanks are
not used. If multiple tanks are used, the relay must be the first one in order
}
