% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{time_at_depth}
\alias{time_at_depth}
\title{time at depth}
\usage{
time_at_depth(dive, depth)
}
\arguments{
\item{dive}{\code{\link[DiveR]{dive}} object.}

\item{depth}{positive numeric value in meter}
}
\description{
Find the time for a given depth and dive.
}
\examples{
time_at_depth(dive(20,40), 3)
time_at_depth(dive(20,40), 10)
time_at_depth(dive(20,40), 20)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
