% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{times_inf}
\alias{times_inf}
\title{times_inf}
\usage{
times_inf(x, col)
}
\arguments{
\item{x}{a \code{\link[mn90]{dive}} object.}

\item{col}{a color value}
}
\description{
times_inf
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
