% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_conso.R
\name{volume}
\alias{volume}
\alias{volume.tank}
\title{volume
\code{volume} retrieve the volume of tank(s).}
\usage{
volume(object)

\method{volume}{tank}(object)
}
\arguments{
\item{object}{is a DiveR object. There is methods for 
\code{\link[DiveR]{tank}} object.}
}
\value{
single numeric value for \code{\link[DiveR]{tank}} object.
 
 @examples 
 t <- tank(12,200)
 volume(object = t)
}
\description{
volume
\code{volume} retrieve the volume of tank(s).
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
