% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_utils.R
\name{add_desat}
\alias{add_desat}
\title{add_desat}
\usage{
add_desat(dtcurve, desat, ascent_speed = 10, secu = FALSE)
}
\arguments{
\item{dtcurve}{a depth time curve in a data.frame with 2 columns depth and 
time. Depths values are meters (positive values) and time is in minute.}

\item{desat}{a desat object that contain the desaturation stops of given dive}

\item{ascent_speed}{Ascent_speed in meter/minute. 10 m/min by default. 
Most dive table advice to limite this speed to 20M/min maximum.}

\item{secu}{security decompression stop of 3 min at 3 m. FALSE by default.}
}
\description{
Add the desaturation stops to the dive curve. It also check if there is no
depth point higher and before the stops, to prevent accidents.
}
