% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn90.R
\name{check_val}
\alias{check_val}
\title{check_val}
\usage{
check_val(val, zero = FALSE)
}
\arguments{
\item{val}{a single positive numeric value}
}
\description{
custom function for checking if values are single positive numeric.
}
