% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conso.R
\name{conso}
\alias{conso}
\title{conso}
\usage{
conso(dive, bloc, cons = 20, mid = 100, reserve = 50, rule = NULL)
}
\arguments{
\item{dive}{\code{\link[DiveR]{dive}} object}
}
\value{
conso, a conso class object.
}
\description{
conso
}
\details{
See \code{\link[DiveR]{tablecheck}} for limit values of depth and time.
}
\examples{
c1 = conso(dive = dive(20,40), bloc = bloc(12, 230), cons = 20, mid = 100, reserve = 50)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
