% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{cust_axis}
\alias{cust_axis}
\title{cust_axis}
\usage{
cust_axis(dive, col, shift = NULL)
}
\arguments{
\item{dive}{a \code{\link[DiveR]{dive}} object.}

\item{col}{a color}

\item{shift}{a numeric to shift the axis in plots. 
Is used with minutes values.}
}
\description{
Draw the xaxis with time input for the plot.dive function
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
