% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{depth}
\alias{depth}
\alias{depth.dive}
\alias{depth.ndive}
\title{depth}
\usage{
depth(object)

\method{depth}{dive}(object)

\method{depth}{ndive}(object)
}
\arguments{
\item{object}{is a mn90 object. There are methods for 
\code{\link[mn90]{dive}} and \code{\link[mn90]{ndive}} objects.}
}
\value{
It returns a numeric with the depth of the dive. 
Is a vector if working on ndive object
}
\description{
\code{depth} retrieve the depth of a singular or multiple dive sequence.
}
\examples{
# Simple dive
depth(dive(20,40))
# Multiple dives
depth(ndive(dive(20,40), dive(15, 80), inter = 540))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
