% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{depth}
\alias{depth}
\alias{depth.dive}
\title{depth}
\usage{
depth(object)

## S3 method for class 'dive'
# depth(dive)
}
\arguments{
\item{object}{is a mn90 object. There are methods for 
\code{\link[mn90]{dive}} objects.}
}
\value{
It returns a numeric with the duration of the ascent and deco.
}
\description{
\code{depth} retrive the depth of a singular or multiple dive sequence.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
