% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_conso.R, R/methods_dive.R
\name{depth_at_time.conso}
\alias{depth_at_time.conso}
\alias{depth_at_time}
\alias{depth_at_time.dive}
\title{depth at time}
\usage{
\method{depth_at_time}{conso}(object, time)

depth_at_time(object, time)

\method{depth_at_time}{dive}(object, time)
}
\arguments{
\item{object}{\code{\link[DiveR]{dive}} object.}

\item{time}{positive numeric value in minute}
}
\description{
Find the depth for a given time and \code{\link[DiveR]{dive}}. 
Also work on conso
}
\examples{
depth_at_time(dive(20,40), 38)
depth_at_time(dive(20,40), 41)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
