% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{depth_at_time}
\alias{depth_at_time}
\title{depth at time}
\usage{
depth_at_time(dive, time)
}
\arguments{
\item{dive}{}

\item{time}{positive numeric value in minute}
}
\description{
Transform minute variable into a character string in hour format.
}
\examples{
depth_at_time(dive(20,40), 38)
depth_at_time(dive(20,40), 41)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
