% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{depths_inf}
\alias{depths_inf}
\title{depths_inf}
\usage{
depths_inf(x, col, only_pal = FALSE)
}
\arguments{
\item{x}{a \code{\link[mn90]{dive}} object.}

\item{col}{a color value}

\item{only_pal}{set to \code{FALSE} by default,}
}
\description{
depths_inf
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
