% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_utils.R
\name{depths_inf}
\alias{depths_inf}
\alias{times_inf}
\title{depths_inf}
\usage{
depths_inf(x, col = "black", only_desat = FALSE)

times_inf(x, col = "black")
}
\arguments{
\item{x}{a \code{\link[DiveR]{dive}} object.}

\item{col}{a color value}

\item{only_desat}{set to \code{FALSE} by default,}
}
\description{
Create information about depth and time to use with text function 
in plot.dive. 
Find the places and labels to show depending on the dive curve.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
