% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_table.R
\name{desat_table}
\alias{desat_table}
\title{desat_table}
\usage{
desat_table(dtcurve, maj = 0)
}
\arguments{
\item{dtcurve}{a depth time curve in a data.frame with 2 columns depth and 
time. Depths values are meters (positive values) and time is in minute.}

\item{maj}{majoration time in minute in case of consecutive dive. 
0 by default.}
}
\description{
desat_table
}
