% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_table.R
\name{desat_table}
\alias{desat_table}
\title{desat_table}
\usage{
desat_table(dtcurve, maj = 0, altitude = 0, ppo2 = 0.209)
}
\arguments{
\item{dtcurve}{a depth time curve in a data.frame with 2 columns depth and 
time. Depths values are meters (positive values) and time is in minute.}

\item{maj}{majoration time in minute in case of consecutive dive. 
0 by default.}

\item{altitude}{Altitude of the dive in meter. Default is 0 m (sea level).}

\item{ppo2}{Partial pressure of dioxygen in bar. Default is 0.209 bar}
}
\value{
a desat object, which is a list with a data.frame containing 
desaturation stops at 9, 6 and 3 m depth. Next element is the dive group
for possible second dive and lastly the times at which the desaturation
stops occur during the dive. The last element is "table" because it's made with 
tables. TODO : this has been modified with refacto !
}
\description{
Extract time for desat stop using the MN90 tables. Desat stop expected are
at 3, 6 and 9 meters depths.
}
\details{
Dive time used in table is the maximum time in the dtcurve table, exept the 
last one. This is important as the last time value is expected to be the 
time at which the diver reach surface.
}
