% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiveR.R
\name{dive}
\alias{dive}
\title{dive}
\usage{
dive(
  depth = 20,
  time = 40,
  secu = TRUE,
  ascent_speed = 10,
  maj = 0,
  desat_model = c("table"),
  hour = NULL,
  dist = NULL,
  speed = NULL,
  way = c("OW", "WB")
)
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.
A single value is needed for square dive, however if a vector is provided
a decompression model need to be selected. Default is 20 meter}

\item{time}{Duration of the dive in minute. Need to be positive values.
A single value is needed for square dive, however if a vector is provided
a decompression model need to be selected. Default is 40 minutes}

\item{secu}{security decompression stop of 3 min at 3 m. FALSE by default.}

\item{ascent_speed}{Ascent_speed in meter/minute. 10 m/min by default. 
Most dive table advice to limite this speed to 20M/min maximum.}

\item{maj}{Time majoration for the dive. 
Only used by table decompression model.}

\item{desat_model}{Which desaturation model is used to compute desaturation
stops during ascent, to eliminate nitrogen. Default is tables as only
this one works today.}

\item{hour}{NULL not implemented yet}

\item{dist}{a distance vector}

\item{speed}{speed of the diver}

\item{way}{If the dive is one way (one way : 'OW') or if the diver return by 
the same depth (way back : 'WB').}
}
\value{
dive, a dive class object.
}
\description{
dive
}
\details{
See \code{\link[DiveR]{tablecheck}} for limit values of depth and time.
}
\examples{
dive = dive(depth = 39, time = 22, secu = TRUE, ascent_speed = 10)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
