% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn90.R
\name{dive}
\alias{dive}
\title{dive}
\usage{
dive(depth = 20, time = 40, secu = TRUE, vup = 10, maj = 0, hour = NULL)
}
\arguments{
\item{depth}{in meter}

\item{time}{in minute}

\item{secu}{true by default, secu deco stage 3 min at 3 meter}

\item{vup}{10 m/min by default}

\item{maj}{0 by default}

\item{hour}{NULL not implemented yet}
}
\value{
dive, a dive class object.
}
\description{
dive
}
\details{
See \code{\link[mn90]{tablecheck}} for limit values of depth and time.
}
\examples{
dive = dive(depth = 39, time = 22, secu = TRUE, vup = 10)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
