% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiveR.R
\name{dive}
\alias{dive}
\title{dive}
\usage{
dive(
  depth = 20,
  time = 40,
  secu = TRUE,
  ascent_speed = 10,
  maj = 0,
  altitude = 0,
  desat_model = c("table", "other"),
  hour = 0,
  way = c("OW", "WB"),
  gas = "AIR"
)
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.
A single value is needed for square dive, however if a vector is provided
a decompression model need to be selected. Default is 20 meter}

\item{time}{Duration of the dive in minute. Need to be positive values.
A single value is needed for square dive, however if a vector is provided
a decompression model need to be selected. Default is 40 minutes}

\item{secu}{security decompression stop of 3 min at 3 m. FALSE by default.}

\item{ascent_speed}{Ascent_speed in meter/minute. 10 m/min by default. 
Most dive table advice to limite this speed to 20M/min maximum.}

\item{maj}{Time majoration for the dive. 
Only used by table decompression model.}

\item{altitude}{Heigth in meter from sea level, it will impact desaturation
process and ascen_speed.. Default is sea level (0m).}

\item{desat_model}{Which desaturation model is used to compute desaturation
stops during ascent, to eliminate nitrogen. Default is tables as only
this one works today.}

\item{hour}{The first immersion hour in minute. Need to be 24h format 
converted in minutes (hour = hours * 60 + minutes). 0 by default.}

\item{way}{If the dive is one way (one way : 'OW') or if the diver return by 
the same depth (way back : 'WB').}

\item{gas}{The gas used by the diver, will be used to modify saturation
computation. Default is 'AIR' but nitrox is possible and can be written as
'NXy' where y is the O2 percentage. Example 'NX32' is a 32% nitrogen tank.
Can go up to 'NX100' and be aware that this will limit the depth due to 
oxygen toxicity wuth a partial pressure of 1.6 bar.}
}
\value{
dive, a list containing a depth/time curve in a data.frame,
the desaturation stops acording to the model used, and some information
about depth to surface time, and input parameters as ascent_speed and secu.
}
\description{
Create a dive class object simulating a dive for given depth and times.
To do so, it create a depth/time curve and find the desaturation stops
according to a desaturation model.
}
\details{
See \code{\link[DiveR]{tablecheck}} for limit values of depth and time.
}
\examples{
dive = dive(depth = 39, time = 22, secu = TRUE, ascent_speed = 10)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
