% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco.R
\name{dtcurve}
\alias{dtcurve}
\title{dtcurve}
\usage{
dtcurve(
  time,
  depth,
  palier,
  ascent_speed = 10,
  dist = NULL,
  speed = NULL,
  way = c("AR", "AS")
)
}
\arguments{
\item{time}{in minute. Max is 180' see Detail !}

\item{depth}{in meter. Max is 65m see Detail !}

\item{palier}{a object palier computed with the palier function. 
Is a list of depth and time for every deco stage.}

\item{ascent_speed}{10 m/min max speed up to the deco. 
Speed between deco is fixed to 6}

\item{dist}{a distance vector}

\item{speed}{speed of the diver}

\item{way}{if the dive is one way or the diver return by the same depth}
}
\value{
a list, with a vector of depths, a vector of time and the dtr value.
}
\description{
Trace a curve of depth and time for given parameters and palier information
}
\details{
This function will stop if the depth > 65 or time > 180 because the table 
are limited to this extent.
However for lower values the table can return NA values.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
