% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{dtime}
\alias{dtime}
\alias{dtime.dive}
\alias{dtime.ndive}
\title{dtime}
\usage{
dtime(object)

\method{dtime}{dive}(object)

\method{dtime}{ndive}(object)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} objects.}
}
\value{
a single numeric value
}
\description{
\code{dtime} retrieve the depth time of a singular or multiple dive sequence.
}
\examples{
# Simple dive
dtime(dive(20,40))
# Multiple dives
dtime(ndive(dive(20,40), dive(15, 80), inter = 540))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
