% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{dtr}
\alias{dtr}
\alias{dtr.dive}
\alias{dtr.palier}
\title{dtr}
\usage{
dtr(object, ...)

\method{dtr}{dive}(object, ...)

\method{dtr}{palier}(object, ..., vup = 10)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} and \code{\link[DiveR]{palier}} objects.}

\item{...}{other arguments passed to the dtr.palier method.
\describe{
  \item{depth}{a numeric in meter}
  \item{vup}{10 meter/minute by default.}
}}

\item{vup}{speed of ascent, 10 m/min by default}
}
\value{
It returns a numeric with the duration of the ascent and deco.
}
\description{
\code{dtr} compute the ascent time from a depth and between deco stages 
of a dive.
It can be retrieved from a dive object or computed with a depth and 
a palier object.
}
\examples{
dtr(dive(20,40))
# absurd slowness with vup = 1
dtr(dive(20,40, vup = 1))
# from a palier object
dtr(palier(20, 40), depth = 20, vup = 10)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
