% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{dtr}
\alias{dtr}
\alias{dtr.dive}
\alias{dtr.palier}
\title{dtr}
\usage{
dtr(object, ...)

## S3 method for class 'dive'
dtr(object = dive)

## S3 method for class 'palier'
dtr(object = palier, depth, vup)
}
\arguments{
\item{object}{is a mn90 object. There are methods for \code{\link[mn90]{dive}} and \code{\link[mn90]{palier}} objects.}

\item{...}{other arguments passed to or from other methods}

\item{vup}{10 meter/minute by default.}

\item{depth}{a numeric in meter}
}
\value{
It returns a numeric with the duration of the ascent and deco.
}
\description{
\code{dtr} returns the ascent and deco time of a dive.
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
