% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dive.R
\name{dtr}
\alias{dtr}
\alias{dtr.dive}
\title{dtr}
\usage{
dtr(object, ...)

\method{dtr}{dive}(object, ...)
}
\arguments{
\item{object}{is a DiveR object. This method is defined for 
\code{\link[DiveR]{dive}} objects.}
}
\value{
It returns a numeric with the duration of the ascent 
to end of the dive.
}
\description{
\code{dtr} retrieve time from end dive time to the surface including
desaturation time
}
\examples{
dtr(dive(20,40))
# absurd slowness with ascent_speed = 1
dtr(dive(20,40, ascent_speed = 1))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
