% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{grp}
\alias{grp}
\title{Deco Group}
\format{An object of class \code{data.frame} with 25 rows and 27 columns.}
\source{
Trucco, Jean-Noël; Biard, Jef; Redureau, Jean-Yves; Fauvel, Yvon (3 May 1999). 
pdf in french :
\href{http://renaud.helstroffer.free.fr/sub_technique/fiche_initiation/mn90.pdf}{"Table Marine National 90 (MN90): Version du 03/05/1999"}
}
\usage{
grp
}
\description{
A dataset containing an data.frame with the group letter depending on depth 
and time.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
\keyword{datasets}
