% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_table.R
\name{majoration}
\alias{majoration}
\title{majoration}
\usage{
majoration(depth, group = "Z", inter = 16)
}
\arguments{
\item{depth}{depth in meter. Must be a single positive value 
with a maximum is 60m.}

\item{group}{byt default "Z", the deco group indicated by a letter. 
This value is indicated in a desat object computed with the desat_table 
function.}

\item{inter}{surface interval time in minute. Must be a single positive value 
with a minimum of 16 minutes (default). Above 720 (12h), the function will
always return 0.}
}
\value{
a time value to add to the second dive time for using tables.
}
\description{
Compute the time majoration to a second dive at a specific depth. 
Is related to the residual nitrogen time.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
