% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn90.R
\name{ndive}
\alias{ndive}
\title{ndive}
\usage{
ndive(dive1, dive2, inter = 16)
}
\arguments{
\item{dive1}{the first dive, obtained by the dive function}

\item{dive2}{the first dive, obtained by the dive function. This one will be
modified with a majoration obtained from dive1 and the interval.}

\item{inter}{16 by default, interval between dives in minutes}
}
\value{
ndive, a ndive class object.
}
\description{
ndive
}
\details{
See \code{\link[mn90]{tablecheck}} for limit values of depth and time 
of a dive.
}
\examples{
dive1 = dive(depth = 39, time = 22, secu = TRUE, vup = 10)
dive2 = dive(depth = 20, time = 40, secu = TRUE, vup = 10)
divet = ndive(dive1, dive2, inter = 30)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
