% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrox.R
\name{nitrox_depth}
\alias{nitrox_depth}
\title{nitrox_depth}
\usage{
nitrox_depth(depth, ppn2 = 0.791)
}
\arguments{
\item{depth}{Depth of the dive in meter. Need to be positive values.}

\item{ppn2}{Partial pressure of nitrogen in bar. Default is 0.791 bar}
}
\value{
The equivalent depth in meter when using this nitrox partial pressure
in a tank.
}
\description{
Compute equivalent depth when using this nitrox partial pressure
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
