% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{plot.dive}
\alias{plot.dive}
\title{plot.dive}
\usage{
\method{plot}{dive}(
  x,
  ...,
  dtr_print = TRUE,
  hour_print = TRUE,
  line_print = TRUE,
  deco_print = TRUE,
  depth_print = TRUE,
  time_print = TRUE,
  def_cols = FALSE,
  legend = FALSE,
  add = FALSE
)
}
\arguments{
\item{x}{an object of class dive.}

\item{...}{every argument for the \code{\link[graphics]{plot}} function 
such as graphical parameters for lines.
Classical \code{\link[graphics]{graphical parameter}} applies on the 
dive curve.
\itemize{
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{xlab} set to \code{'Time'} by default, 
  see \code{\link[graphics]{title}}.
  \item \strong{ylab} set to \code{'Depth'} by default, 
  see \code{\link[graphics]{title}}.
  }}

\item{dtr_print}{set to \code{TRUE} by default, whether there is text and 
line for ascent time.}

\item{hour_print}{set to \code{TRUE} by default, whether there is the hour
at beginning and end of the dive.}

\item{line_print}{set to \code{TRUE} by default, whether there is lines 
for depths.}

\item{deco_print}{set to \code{TRUE} by default, whether there is information
about desat stop for depths and times.}

\item{depth_print}{set to \code{TRUE} by default, whether there is depth 
values along the dive.}

\item{time_print}{set to \code{TRUE} by default, whether there is time 
values along the dive.}

\item{def_cols}{set to \code{FALSE} by default, modify plot colors to match
shinyapp css.}

\item{legend}{set to \code{FALSE} by default, print legend about the ascent 
speeds.}

\item{add}{set to \code{FALSE} by default, to add another dive plot 
on a precedent one.}
}
\description{
Plot the dive curve depending on time and depth. 
Only represent sqare profile
}
\seealso{
\itemize{
\item \code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and 
\code{\link[graphics]{par}} for plot parameter that were omitted 
on this documentation
\item \code{\link[mn90]{dive}} for every aspect about dive object creations.
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
