% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{plot.ndive}
\alias{plot.ndive}
\title{plot.ndive}
\usage{
\method{plot}{ndive}(x, ..., text_print = TRUE, add = FALSE)
}
\arguments{
\item{x}{an object of class ndive.}

\item{...}{every argument for the \code{\link[graphics]{plot}} function 
such as graphical parameters for lines.
Classical \code{\link[graphics]{graphical parameter}} applies on the 
dive curve
\itemize{
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{xlab} set to \code{'Time'} by default, 
  see \code{\link[graphics]{title}}.
  \item \strong{ylab} set to \code{'Depth'} by default, 
  see \code{\link[graphics]{title}}.
  }}

\item{text_print}{set to \code{TRUE} by default, whether there is a text 
for depths and duration of deco stops.}

\item{add}{set to \code{FALSE} by default, to add another dive plot 
on a precedent one.}
}
\description{
Plot the dive curve depending on time and depth. 
Only represent sqare profile
}
\seealso{
\itemize{
\item \code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and 
\code{\link[graphics]{par}} for plot parameter that were omitted 
on this documentation
\item \code{\link[mn90]{ndive}} for every aspect about 
ndive object creations.
}
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
