% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{secu}
\alias{secu}
\alias{secu.dive}
\title{secu}
\usage{
secu(object)

\method{secu}{dive}(object)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} objects.}
}
\value{
Boolean
}
\description{
\code{secu} retrieve if a dive was set with secu = TRUE or FALSE.
}
\examples{
# Without desat stop
secu(dive(20,40, secu = TRUE))
secu(dive(20,40, secu = FALSE))
# With desat stop
secu(dive(20,60, secu = TRUE))
secu(dive(20,60, secu = FALSE))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
