% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_code.R
\name{speed}
\alias{speed}
\alias{speed.dive}
\title{speed}
\usage{
speed(object)

\method{speed}{dive}(object)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} objects.}
}
\value{
It returns a numeric with the duration of the ascent and deco.
}
\description{
\code{speed} compute the ascent speed from a depth and between deco stages 
of a dive.
It can be retrieved from a dive object.
}
\examples{
# Dive with default parameters
speed(dive(20,40, ascent_speed = 10, secu = TRUE))
# different ascent_speed
speed(dive(20,40, ascent_speed = 15, secu = TRUE))
# No desat stop induce NA for plt speed (between desat stop speed)
speed(dive(20,40, ascent_speed = 10, secu = FALSE))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
