% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{summary.dive}
\alias{summary.dive}
\alias{summary.palier}
\title{summary}
\usage{
\method{summary}{dive}(object, ...)

\method{summary}{palier}(object, ...)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}} and \code{\link[DiveR]{palier}} objects.}

\item{...}{other arguments passed to the dtr.palier method.
\describe{
  \item{plot}{FALSE by default}
}}
}
\value{
A brief summary of a dive and its main parameters.
}
\description{
summary of a dive object
}
\examples{
summary(palier(depth = 20, time = 50))
summary(palier(depth = 20, time = 40,secu = FALSE))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
