% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{summary.dive}
\alias{summary.dive}
\alias{summary.palier}
\title{summary}
\usage{
\method{summary}{dive}(object, ...)

## S3 method for class 'palier'
# summary(palier)
}
\arguments{
\item{object}{is a mn90 object. There are methods for 
\code{\link[mn90]{dive}} and \code{\link[mn90]{palier}} objects.}

\item{...}{other arguments passed to the dtr.palier method.
\describe{
  \item{plot}{FALSE by default}
}}
}
\value{
A brief summary of a dive and its main parameters.
}
\description{
summary of a dive object
}
\examples{
summary(palier(depth = 20, time = 50))
summary(palier(depth = 20, time = 40,secu = FALSE))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
