% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R, R/methods_conso.R
\name{summary.dive}
\alias{summary.dive}
\alias{summary.palier}
\alias{summary.tank}
\alias{summary.conso}
\title{summary}
\usage{
\method{summary}{dive}(object, ...)

\method{summary}{palier}(object, ...)

\method{summary}{tank}(object, ...)

\method{summary}{conso}(object, ...)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}}, \code{\link[DiveR]{tank}} and \code{\link[DiveR]{palier}} objects.}

\item{...}{other arguments passed to the dtr.palier method.
\describe{
  \item{plot}{FALSE by default}
}}
}
\value{
A brief summary of a dive and its main parameters.
}
\description{
summary of a dive object
}
\examples{
summary(palier(depth = 20, time = 50))
summary(palier(depth = 20, time = 40,secu = FALSE))

summary(tank(vol = 12, press = 200))

Tank_15L <- tank(vol = 15, press = 200)
viable <- conso(dive = dive(20,40), tank = Tank_15L, 
                cons = 20, failure_label = 'Air failure')
summary(viable)

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
