% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_conso.R, R/methods_dive.R
\name{summary.tank}
\alias{summary.tank}
\alias{summary.conso}
\alias{summary.dive}
\alias{summary.desat}
\title{summary}
\usage{
\method{summary}{tank}(object, ...)

\method{summary}{conso}(object, ...)

\method{summary}{dive}(object, ...)

\method{summary}{desat}(object, ...)
}
\arguments{
\item{object}{is a DiveR object. There are methods for 
\code{\link[DiveR]{dive}}, \code{\link[DiveR]{tank}} and
 \code{\link[DiveR]{conso}} objects.}

\item{...}{other arguments not used}
}
\value{
A brief summary of a dive and its main parameters. 

summary(dive(20, 40, secu = FALSE))
summary(dive(20, 40, secu = TRUE))
summary(dive(20, 45))
summary(dive(39, 22))
summary(dive(50, 22))
}
\description{
summary method for class \code{\link[DiveR]{dive}}.
}
\examples{
summary(tank(vol = 12, press = 200))

Tank_15L <- tank(vol = 15, press = 200)
viable <- conso(dive = dive(20,40), tank = Tank_15L, 
                cons = 20, failure_label = 'Air failure')
summary(viable)

summary(desat_table(init_dtcurve(20, 40)))
summary(desat_table(init_dtcurve(20, 45)))
summary(desat_table(init_dtcurve(39, 22)))
summary(desat_table(init_dtcurve(50, 22)))

}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
