% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desat_table.R
\name{tablecheck}
\alias{tablecheck}
\alias{max_depth_time}
\title{tablecheck}
\usage{
tablecheck(depth, time, force = FALSE)

max_depth_time(depth, force = FALSE, no_deco = FALSE)
}
\arguments{
\item{depth}{depth in meter. Must be a single positive value 
with a maximum is 65m.}

\item{time}{time in minute. Must be a single positive value 
with a maximum is 180 min}

\item{force}{FALSE by default, if TRUE don't stop the function but 
return a TRUE/FALSE value}

\item{no_deco}{FALSE by default, if TRUE return the time without deco}
}
\value{
Nothing expect when \code{force} parameter is TRUE. 
Is designed to stop code and returns errors

Single numeric value, the max time possible to dive at the given 
depth is the MN90 table.
}
\description{
Check if the parameters are possible with table information.

Max time present in the table for a given depth, with or without deco
}
\details{
This function will stop if the depth > 65 or time > 180 because the table 
are limited to this extent (the actual table in dataset can evolve)
However for lower values the table can return NA values. 
This NA return is avoided in the shinyapp.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
