% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco.R
\name{tablecheck}
\alias{tablecheck}
\alias{max_depth_t}
\title{tablecheck}
\usage{
tablecheck(depth, time, force = FALSE)

max_depth_t(depth)
}
\arguments{
\item{depth}{in meter. Max is 65m see Detail !}

\item{time}{in minute. Max is 180' see Detail !}

\item{force}{FALSE by default, if TRUE don't stop the function but 
return a TRUE/FALSE value}
}
\value{
a single numeric value, the max time possible to dive at the given 
depth is the MN90 table.
}
\description{
Check if the parameters are possible with table information.

Max time present in the table for a given depth.
}
\details{
This function will stop if the depth > 65 or time > 180 because the table 
are limited to this extent (the actual table in dataset can evolve)
However for lower values the table can return NA values. 
This NA return is avoided in the shinyapp.
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
